refactor: NavBarMain 전체적인 코드 분리
