+++ src/js/components/header/NavBarMain/searchbar/View.js
@@ -1,0 +1,9 @@
+import { searchBarModal } from "./template.js";
+
+export class SearchBarView {
+  constructor() {}
+
+  renderSearchBarModalElement(state) {
+    return `${searchBarModal.template(state)}`;
+  }
+}
